+++ config.go
@@ -2,6 +2,17 @@ package noaa
 
 import "strings"
 
+// deprecated
+// Default values for the weather.gov REST API config which will
+// be replaced by Config. These are subject to deletion in the future.
+// Instead, use noaa.GetConfig followed by:
+//     Config.BaseURL, Config.UserAgent, Config.Accept
+const (
+	API       = "https://api.weather.gov"
+	APIKey    = "github.com/icodealot/noaa" // User-Agent default value
+	APIAccept = "application/ld+json"       // Changes may affect struct mappings below
+)
+
 // Config instance for the API calls executed by the NOAA client.
 var config = GetDefaultConfig()
